[iOS][updates] Migrate tests to Swift Testing (2/3)
