+++ packages/expo-updates/ios/Tests/AppLauncherWithDatabaseTests.swift
@@ -40,6 +40,7 @@ class AppLauncherWithDatabaseMock: AppLauncherWithDatabase {
 }
 
 @Suite("AppLauncherWithDatabase", .serialized)
+@MainActor
 class AppLauncherWithDatabaseTests {
   var testDatabaseDir: URL
   var db: UpdatesDatabase
